TCXR_HW/API_Python/requirements.txt
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
